+++ locales/en.lua
@@ -48,6 +48,8 @@ return {
     ['hud.money_gain']   = '+$%d',
     ['hud.health_loss']  = '-%d',
     ['hud.ammo']         = '%s  Ammo: %d/%d',
+    ['hud.medkit']       = 'MED KIT  x%d',
+    ['hud.medkit_hint']  = 'CLICK: HEAL +%d HP',
     ['hud.gun_pickup']   = 'Press E to pick up %s',
     ['hud.door_open']    = 'Press E to open - $%d',
     ['hud.door_locked']  = 'Locked - need $%d',